LayoutTests/platform/glib/imported/w3c/web-platform-tests/fetch/private-network-access/redirect.https.window-expected.txt


PASS local to private to public: success.
PASS local to private to local: success.
FAIL private to private to local: failed preflight. assert_equals: error expected (string) "TypeError: Failed to fetch" but got (undefined) undefined
FAIL private to private to local: success. assert_equals: error expected (undefined) undefined but got (string) "TypeError: Load failed"
PASS private to private to local: no-cors success.
FAIL private to local to private: failed preflight. assert_equals: error expected "TypeError: Failed to fetch" but got "TypeError: Load failed"
FAIL private to local to private: success. assert_equals: error expected (undefined) undefined but got (string) "TypeError: Load failed"
PASS private to local to private: no-cors success.
FAIL public to private to local: failed first preflight. assert_equals: error expected "TypeError: Failed to fetch" but got "TypeError: Load failed"
FAIL public to private to local: failed second preflight. assert_equals: error expected "TypeError: Failed to fetch" but got "TypeError: Load failed"
FAIL public to private to local: success. assert_equals: error expected (undefined) undefined but got (string) "TypeError: Load failed"
PASS public to private to local: no-cors success.
